# api.yml
openapi: 3.0.1
info:
  title: Text service
  contact:
    email: open-community@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.php
  version: 1.0.0
servers:
  - description: "Localhost server"
    url: http://localhost:3000
...
paths:
  /texts:
    get:
      summary: "Return a list of texts"
      operationId: listTexts
      tags:
        - text
      parameters:
        - name: author.account.id
          description: "Account ID of the text authors: If defined, then will return each text written by each of these authors"
          in: query
          schema:
            type: array
            items:
              type: string
              format: base64-url
              
        - name: author.identity.id
          description: "Identity ID of the text authors: If defined, then will return each text written by each of these authors"
          in: query
          schema:
            type: array
            items:
              type: string
              format: base64-url
              
        - name: content
          in: query
          description: "If defined, will return only text containing the given search in the content"
          schema:
            type: array
            items:
              type: string
              format: url-encoded
              
        - name: context.id
          description: "Context ID of the text"
          in: query
          schema:
            type: array
            items:
              type: string
              format: base64-url

        - name:  creation-date.max
          description: "Will return only texts created at this date or before (lower or equal)"
          in: query
          example: "2019-01-31T23:59:59Z"
          schema:
            type: string
            format: "ISO 8601 Extended Format"

        - name: creation-date.min
          description: "Will return only texts created at this date or after (greater or equal)"
          in: query
          example: "2019-01-01T00:00:00Z"
          schema:
            type: string
            format: "ISO 8601 Extended Format"
          
        - name: deletion-date.max
          description: "Will return only texts deleted at this date or before (lower or equal)"
          in: query
          example: "2019-02-10T23:59:59Z"
          schema:
            type: string
            format: "ISO 8601 Extended Format"
            
        - name: deletion-date.min
          description: "Will return only texts deleted at this date or after (greater or equal)"
          in: query
          example: "2019-01-02T00:00:00Z"
          schema:
            type: string
            format: "ISO 8601 Extended Format"

        - name: id
          description: "ID of all identities to retrieve"
          in: query
          schema:
            type: array
            items:
              type: string
              format: base64-url
              
        - name: pagination.offset
          description: "If defined, the search will skip the first \"pagination.offset\" result and return the next ones."
          in: query
          schema:
            default: 0
            type: number
            
        - name: pagination.size
          description: "If defined, only the \"pagination.size\" first results after offset will be returned. By default, all results will be returned"
          in: query
          schema:
            type: number

        - name: search
          description: "If defined, will return only text containing the given search in the title or content"
          in: query
          schema:
            type: array
            items:
              type: string
              format: url-encoded
              
        - name: sort
          description: > 
            Enable result sorting.
          in: query
          schema:
            default: 
              - _score
            type: array
            items:
              type: string
              enum:
                - +creation-date
                - -creation-date
                - +deletion-date
                - -deletion-date
                - +content
                - -content
                - +title
                - -title
                - _score
          
        - name: owner.id
          description: "ID of all text owners"
          in: query
          schema:
            type: array
            items:
              type: string
              format: base64-url
          
        - name: title
          in: query
          description: "If defined, will return only text containing the given search in the title"
          schema:
            type: array
            items:
              type: string
              format: url-encoded
          
      responses:
        200:
          description: List of all texts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Text"
...
  /text/{id}/authors:
    get:
      summary: List all authors of a text
      operationId: listAuthors
      tags:
        - accounts
        - authors
        - identities
      parameters:
        - name: id
          in: path
          description: "ID of the text"
          required: true
          schema:
            type: string
            format: base64-url
            
      responses:
        200:
          description: "List of authors"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountIdentity"
                
        404:
          description: "Text not found"
      
    put:
      summary: Declare one or several new authors of an text.
      operationId: addAuthors
      tags:
        - authors
      parameters:
        - name: id
          in: path
          description: "ID of the text"
          required: true
          schema:
            type: string
            format: base64-url
            
      requestBody:
        description: "Authors to add"
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AccountIdentity"
                
      responses:
        200:
          description: "Authors added"
          
        404:
          description: "Text not found"
...
components:
  schemas:
    Text:
      type: object
      properties:
        id:
          type: string
          description: ID of the text
          
        text:
          type: string
          description: "Content of the text"
          
        formatting:
          type: object
          description: "Formatting data"

        owners:
          type: array
          description: "List of text owners"
          items:
            type: string
            format: base64-url

        authors:
          type: array
          description: "List of text authors"
          items:
            $ref: "#/components/schemas/AccountIdentity"
    AccountIdentity:
      type: object
      properties:
        account:
          type: object
          properties:
            id: 
              type: string
              description: ID of the owner account
              
        identity:
          type: object
          properties:
            id:
              type: string
              description: ID of the owner identity